Project refactoring for future publishing
